[ADD] 게시글 수정 기능 구현 - 김인엽

+++ GroomHim/src/main/java/com/him/groomhim/question/repository/QuestionRepository.java
@@ -10,5 +10,5 @@ import java.util.List;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
 
-    Question findByQuestionNo(int questionNo);
+    Question findByQuestionNo(Long questionNo);
 }
